Refactor Server interface to support operations with devices
This change is mostly a preparation for the workspace list --all
implementation. It introduces a concept of a project explicitly and
makes the Server interface more granular, so in the future, it will be
possible to put a task system with a do/undo interface. Thus, actions
performed over a workspace's state could be tracked, rolled back and
debugged easier.