Add support for copied project directories
In this change, we introduce checking if the original project directory
and its lock file still exist. If they do, it means the current project
directory was copied and the project-id needs to be recreated. So, all
the previously created instances stay with the old project and the new
one does not get any instances assigned to it.