Bootstrap Workspace API using pebble and snapd
This is a single change that separates workspace into workspace
and workspaced to represent the extraction of the API layer. Mostly,
it uses the Pebble's daemon implementation which, in turn, is based
on snapd's. The workspace client is not changed to work with the
daemon yet and still operates with Overlord directly. The upcoming
changes to the client and server will move all the available commands
to using API calls.