Add /v1/projects implementation for a single project
This change also reworks the project management as unties it
from overlord which it did not belong from the state management
point anyway. Now, project is a separate package that encapsulates
the relationships with the LXD's instance configuration for
the project directories (i.e. updates conf when a bind mount changes,
tracks .lock files integrity and recover)