Add context to all LXD backend requests, rework Project to fit API requests
This is a large change that introduces a context to all the requests that
the client sends to the workspace backend (LXD). The context must have
several fields set: project-id and a user (system user name). Thus, we
support a multi-user story when users can share a project directory and
continue to launch workspaces for the same file. For example, if a user
has .workspace.nimble.yaml in /shared/project, they will be able to launch
their workspace even if one exists that was created by another user in the same
directory. Also, the Project code was largerly reworked to better accomodate
the needs of the API and simplify the object by extracting some of its
responsibilities to a separate set of functions independent of the object itself