Add workspace launch using Workspace API (initial implementation)
This work continued moving Project to the workspacebackend package
and replaced the launch CLI implementation with calls to the API.
The POST /v1/projects/{id}/workspaces was implemented and now
supports executing an action on a list of workspaces (in such a case,
the action will be executed as a single change. Thus, if failed,
it will rollback progress for the entire change where applicable).
Also, the changeset introduces a progress bar borrowed from snap