Add forceful workspace delete option for the undoCreateWorkspace logic
Sometimes, a launch operation can be interrupted during the start process
and when the undo process unrolls, it needs to delete the instace. But
between the interruption time and an attempt to delete the instance, the
instance may have been started already which results in a failed undo
for the change. Thus, we should delete forcefully if we are in our undo
chain