Extract launching multiple workspaces to a separate function
This is something we want to be less exposed to the API layer. Previously,
the response logic loaded workspaces and created launch task sets for them.
It should not be a responsibility of the daemon package. The daemon will only
parse the query and dispatch it to the corresponding managers to handle
the conditional taskset creation to keep the logic strictly in one place