Separate files and workspace that are returned by GetAllWorkspaces
This change makes GetAllWorkspaces of the backend return two
separate sets: files without workspace instances and workspace instances.
It is needed to make WorkspaceProps (i.e. a workspace structure) to be
created only for the actual workspaces that were launched before. This,
in turn, will allow us to introduce additional methods to the workspace
which would not make sense for files-only workspaces (e.g. set state)