Move LinkSdk and UnlinkSdk logic to Workspace from the corresponding tasks
This is the change that encapsulates the logic of work with the SDK records
in the Workspace, thus, the clients do not have to Marshall/Unmarshall JSON
from the LXD configuration. Also, it removes the Config field that was
previous part of the workspace. It is a plain copy of the LXD config which
worked in the first iterations, but now it is time for Workspace to be
rather independent of the LXD structure and conceal its implementation of
the confguration. The new refresh request implementation will use the
Workspace methods to initiate a refresh check with the store