Make WorkspaceBackend to fully load and initialise Workspace
It gets rid of the generic GetWorkspacesByConfig that returned a semi
initialised Workspace structure and was used in places were just a plain
LXD Instance structure sufficed (e.g. for the bind mounts scaninng).
This change made a creation of Workspace explicit. The structure
contains a reference to its file and SDKs. This is required for the
tasks to operate on a workspace without the need to find its associated
files or parse LXD instance json config