Update refresh to use a separate project space instead of renaming
Previously, refresh did a rename of an instance that was about to be
refreshed. Now, the workspace backend introduces a system project that
will be used to keep the workspaces that are being refreshed. It solves
the problem of the old version of the workspace appearing in the list
output during the operation with an Error status