Add WaitStatus support to the state system from snapd
This is a single commit that adds the latest changes (specifically,
WaitStatus support for tasks, as of e3215d6f1f) of the snapd state
system to workspace. This will allow some workspace features to pause a
change in progress, wait for user's input and resume without putting a
workspace in an Error or Undone state. For an example, see workspace
refresh with --continue and --abort flags