Implement wait-on-error with continue and abort using Wait Status
This change uses the previously merged WaitStatus for tasks from snapd.
It makes the implementation free of the TaskRunner hacks that changed
the state of the Change from Error to Doing on continue. Now, if an
error was encoutered, we set the task status to wait and let the user to
resolve the issue. Once fixed we change the task status to Doing and let
the state management do its execution magic without slipping into an
Error state