Report correct workspace state for the refresh operations
Prior to this moment, all the long running operations on a workspace did
not affect its reported state if it was Running. Hence, any list command
would show the workspace as Ready. It is incorrect for two reasons.
First, a user should not be able to run another command that may change
the workspace whilst it is in progress (i.e. Pending). Second, the state
must be reported correctly, i.e. if a workspace is being refreshed, the
list should render it as Pending. This commit does not solve the first
problem yet. A user can still try to run commands over a Pending
workspace. It does solve the state reporting issue (only for the refresh
operations at the moment)