Show correct abort message and fix exit codes
The abort operation should not list the errors that were encountered
during the refresh operation, it only needs to say that the refresh was
aborted to not confuse the user with a list of error messages that they
have seen before. To do that, we will need to analyse the change and
find out if the abort operation has been completed successfully or
produced an error in one of the task's undo logic. If the latter, report
those error messages