Move loading workspaces to be part of the WorkspaceManager
Clean up the api_projects.go logic from knowing too much of the internal
details of the workspace loading and talking to the workspace backend
directly. Instead, the API will now work with the Workspace manager if
it needs to load workspaces. The manager, in turn, will encapsulate the
logic of interacting with the WorkspaceBackend and determiniting of a
workspace state based on the State data