Add missing new line after the group tab
tab and group-tab are reST directives. Each enhancement's given as an
option, written directly below the line containing the directive. The
result is that if any directive gets options added to it in future,
existing documents with text immediately below the directive line will
no longer build. It's not a bad habit to leave a line below directives.