Handle refresh cases with adding a new SDK to the file
The order and quantity of restore and setup hooks depend on the
installed content (i.e. the previous workspace version) and on the new
workspace file. The workspace manager must call save state hooks for
these SDKs that were previously installed and restore hooks for the SDKs
that will be installed