Use stash to preserve and restore previous workspace version
Stashing means putting workspaces away so it becomes invisible to other
commands until a following unstash call. Previously called as 'copy' it
is renamed to use stash in the task and workspace backend interface's
names as something describing the behaviour more precisely. The LXD
backend implements stashing as putting a workspace to a separate LXD
project space called "system project"