Stop an operation if cancelled or when the final task is done
Every task can result in nil, context.Cancelled or an error. We do not
treat the context cancellation as an error as it is an indicator that
the corresponding change was aborted by the client. Instead, we remove
the operation from the list of ongoing operations and let the
cancellation (undo) to execute its logic. Also, if the task resulted in
no error but was marked as the one completing the operation, the
operation will be removed from the list of operations allowing clients
to start another one for the workspace