Add workspace exec and task websocket API
The initial implementation of the exec API that returns a task ID
associated with the command execution. The task ID will be used by the
client to connect to the task's websocket and forward IO of the command
to the user (similarly to how Pebble does it). Currently, the task
websocket API will redirect to the corresponding LXD websocket endpoint