Make task websocket request to fail if exec preparation times out
When a client connects to a task's websocket, the API will wait until
the execution operation is ready and websockets can be connected. If the
execution operation is not ready within 5 seconds, the request will stop
waiting and fail with DeadlineExceeded.

Also, add a command timeout parameter. It is a different sort of timeout
applied to the command execution. The countdown starts from the moment
the execution operation is created in LXD. If exceeded, the command
execution will be cancelled with DeadlineExceeded