Handle HTTP redirects for the task API websocket calls
Previously, it was made as handling a failed websocket handshake and
reading the response to establish a websocket connection with the server
provided in the response's Location. This implementation is clearer as
it probes the task websocket API, gets an http redirect and uses a
websocket dialer to establish the connection. This avoids attempts to
establish a task's websocket connection that we know will fail with bad
handshake anyway due to the redirect