Add workspace remove command
Currently, workspace remove ignores the status of a workspace or whether
it has any operations in progress. This initial version does not account
for a possible refresh operation in pending to clean up corresponding
resources, e.g. stash and volumes. Ideally, it should cancel all the
ongoing changes for the given workspace first and then continue to
remove