Make workspace manager to take care of the actions prereqs
Earlier, the API endpoint used to talk to the Workspace Manager to
validate whether the workspaces in question meet all the required
pre-conditions to execute an action (status, existence, etc). This
change set makes its a responsibility of the manager and the API
endpoint logic would only take care of the input parameters correctness
and dispatching the request to the corresponding function in request.go
of the manager.

Also, this change adds status checking to the workspace remove command.
The removal can only be performed if a workspace is in Ready, Stopped or
Error mode. A workspace that has an operation in progress (i.e. Pending)
will not be removed. This is to be inline with the underlying system of
changes; we cannot reliably interrupt a change that may be in progress
for the given workspace and/or wait for its abortion to finish.
Generally, it does not seem to be a good idea to make one change depend
on another. Thus, the remove (as many others) change is applicable if
there is no other change for the given workspace is in progress