Deduplicate CLI positional args
The deduplication happens on the API side as well; here it is done to
make output to reflect the actual progress correctly. For example, no
duplication about some workspace launched twice if one used it twice in
the command arguments.

Also, a bit more stricter control on the number of allowed positional
args and consistent messaging about the results of a command execution:
all names quoted