Add Plugs and Slots support to the SDK yaml specification
Also, separate the SDK representation into sdk.Setup and sdk.Info. The
former is responsible for the essential SDK information provided by the
store and used in the initial stages of launch, refresh, etc. It is used
to retrieve and install/link an SDK. The latter contains full
information about an SDK that has been installed in a given workspace.
That includes slots, plugs, meta information to facilitate interface
connections work