Add base declaration install, connect and auto-connect policy checker
The policy package acts as a layer to check the plugs and slots
installation, connection and auto-connection subrules. Here we introduce
a policy checker that uses base declaration (currenly containing the
content interface slot declaration only) to allow or permit actions from
these sub-rules.

The SDK install, launch and and other tasks will be using the policy
checker to decide on whether an SDK can be installed (install) and
(auto-)connected (launch, refresh, remove)