Add Interface Manager skeleton to handle connection tasks
Interface manager is responsible for keeping track of the interfaces
connections and handling the logic of the SDKs plug and slot
connections.

The changes to the launch, refresh and remove will follow. These will
introduce new types of tasks to reconnect, disconnect and remove
connections for the workspace's SDKs.

Also, reworked the Projects() method of the workspace LXD backend to
handle scenarios when its client (the interface manager) needs to walk
all the existing projects, workspaces and their SDKs for all users to
reload connections on restart