Separate task data for retrieve and install SDK
The retrieve and install tasks used the same name to exchange
information about the SDK to be installed/downloaded. Separate these, so
now the task responsible for downloading the SDK gets its input from
"sdk-record" and stores its output to "sdk-setup". The latter is
used by the following tasks, e.g. an install task