Add Profile, an interface to manage SDK profiles for workshops
The device interface backend needed a way to update workshop devices
according to the existing interface connections. This change introduces
an SDK profile, an abstraction over a LXD profile. When the interfaces
are connected or disconnected, the backend will create respective
devices (e.g. LXD mounts, desktop unix sockets, GPU, etc), rebuild and
assign the SDK profile to the workshop.