Remove auto-connect task dependency on the retrieval task
The auto-connect should have only the associated SDK name it operates on
as by the execution time it already has the SDK installed in a workshop
and can request the required info from the workshop directly. Hence,
decouple it from the sdk.Setup structure which is only required when an
SDK is installed.

Also, update String() of PlugRef and SlotRef to include the workshop
making is workshop:sdk:plug instead of sdk:plug.