Restore previous SDK connections if refresh failed
workshop refresh can abort at any point and restore the stashed version.
The interface connections and profiles need to be restored together with
the instance. To do so, we initiate the connection task for every SDK
that will be running in a separate lane to not affect the main change
nor other SDK reconnection tasks.

The install candidate policy check was moved to the auto-connect task to
handle scenarios when an SDK is reconnected without linking, i.e. if an
SDK existed in a stashed workshop, it would not be reinstalled and
relinked. Thus, when we restore the SDK interface connections, we must
check the policy again in case anything fiddled with the meta data
content and added some prohibited slots, for example.