Add code review checklist
To adi in self-reviews and keep the incoming changes tidy. Largerly
inspired by M. Fowler's "Tidy First?".

The guide will never intend to be a complete manual that lets a reviewer
to switch into a brain-off mode. This is a set of recommendations that
would provide the biggest returns on investment.