Add workshopctl command
workshopctl uses /v1/workshopctl to access the
available commands from within an SDK hook.

workshopctl will be made available to a workshop
at the time of construction. It requires
WORKSHOP_COOKIES env variable to be set at the
time of call to interact with workshopd and affect
the hook execution logic. The said variable is
always set for any hook type supported by
Workshop.