Add check-health context handler and workshopctl set-health
The context handler consults the "health" field of
the context to determine the health status of an
SDK and choose what to do next. It can choose to
continue the execution normally (okay status),
retry (waiting) or report an error and initiate
the change's abort operation (error).

The "health" property of the context can be set by
an SDK using workshopctl set-health command.