Move workshop health and status under healthstate package
The logic of this change is to make set-health
consistent with already existing statuses and
notes of workshops. In order to achieve that, the
workshop status has been moved under the
healthstate package and a notion of workshop
health has been introduced. The health includes
the status; health  may include a message, code
and timestamp depending on the status.

It decoupled workshop and its status/health
because a workshop's health is calculated based on
the input from set-health, its files integrity and
a possible ongoing operation it may have running
that WorkshopBackend will never be fully aware of.
Therefore, a workshop's health is a concern of the
WorkshopManager and healthstate package.