Pass set-health message and code to the workshop info
The message and code that are provided in a
check-health hook using set-health command should
be shown in the `workshop info` output. For
example, if `workshop info` was called during the
refresh operation and an SDK repored a not okay
health check result, then the corresponding
message and code will be part of the output.

The health check results are not persistent and
they only exist for the duration of the check.
Once the check is over, the SDK is either 'okay'
or 'error'. The former will allow the workshop to
continue the ongoing operation (refresh or
launch). The latter will cause abortion or pause
the operation (in the wait-on-error refresh).

This change has also separated the health state of
a Workshop and a result of a health check for an
SDK. SDKs do not have statuses, only workshops do.
The status is something defined by workshopd and
it cannot be overriden by an external command
directly. However, a user has a control over the
health check result in a check-health hook. The
outcome of the health check for an SDK is
accounted for by workshopd to define the status of
a workshop.