Make workshop info to show the SDK health check status
* `workshop info` will render the results of an
SDK health check as part of its output for the
SDK.

* The healthstate handler resets the retry counter
for the health check to accomodate
non-transactional refresh scenarios, i.e. those
which were run with --wait-on-error and stopped
after an unsuccessful error check. When continued
with --continue, the refresh must restart the
check-health task from the beginning (not
accounting for the previous attempts).