Rework conflicting change detection logic
* Check if there is a conflicting change running
for a specific workshop/project instead of marking
the change's tasks and checking their statuses.
This will be used by all the new operations that
deal with interfaces and can only be executed if a
workshop is Ready.

* Divide the monolitic /v1/projects/{}/workshops/
unit tests into smaller ones to prepare for the
new tests that support workshop remount.