Ensure that plug's dynamic attrs preserved after restart
The new mounts API uses the content interface plug
dynamic attrs to store and retrieve 'source'. This
change updates auto-connect task to ensure that,
after connecting relevant plugs and slots, those
connections's plugs dynamic attrs will be stored
in the state too.

Earlier, if an interface backend added a dynamic
attribute to a plug, it wouldn't have been
preserved as the auto-connections were stored in
the state before the interface backend setup call.