Rework LXD issue workaround to accomodate remount scenarios
When a directory is unmounted, LXD removes the
target path if it is empty. Earlier, Workshop
created a file to make directory preserved. It
makes the remount UX confusing as if a user wants
two workshop to share a source, they won't be able
to do so due to the hidden file in the original
source that prevents the second workshop from
remount.