Rename slot-type and plug-type to *-sdk-type in declarations
Interface policy checks use the type of an SDK to
determine whether it could be installed, connected
or auto-connected. When the interface system was
ported to workshop from snapd it was named as
slot-type and plug-type as in the original version
it was *-snap-type. Now, when the agent SDK is
being introduced, it makes sense to clarify what
entity the type is related to.