Add check for conflicting content plugs targets
Workshop will not be refreshed or launched if
there are SDKs that introduce targets for the
mounts with different sources.

Multiple areas were tried to execute this check.
One of the most logical places would be the
content interface policy / validation. However,
given that interfaces and the repository operate
at the SDK level, it is surprisingly impossible to
validate a candidate plug against plugs provided
by other SDKs at this level. Thus, the check
happens right after an SDK was added to the
Repository. If there is a conflict, the change
will be aborted.