Add end to end tests for workshop disconnect
This test is accompanied some tidy up changes:

* Close the LXD client immediately after a
workshop operation to avoid inactive connections
(these were closed by LXD on timeout later).

* MoveInstance LXD API does not have a stable
behaviour in some cases. E.g. if the workshop was
stashed and then unstashed it could be that an
attempt to get its FS via SFTP will just hang the
client. All attempts to shell in such a container
will just lead to: "Failed to retrieve PID of
executing child process". Another problem would be
that stopping this instance next time will cause a
timeout if not done forcefully. Thus, for every
stash operation, the profile list is set to have
only the default one and the instance is stopped
forcefully.