Store and retrieve sdk.yaml metadata from GcsStore
* Retrieves the sdk.yaml from the store before
initiating launch or refresh for validation
purposes. This is the time when one needs to run
cross-SDK validation, for example, for plug
binding to make sure the plugs from another SDK
exist and are of the right interface.

* Better emulate the actual store behaviour by
giving the store a list of actions for a workshop
(install, refresh an SDK) and getting results back
(a store's SDK info or errors).