Connect bound plugs in interface connection handlers
There are two places where workshop initiates
"connect" tasks: auto-connection tasks on
launch/refresh and an individual connect request.
Both are now aware of the binding which they
obtain from the workshop definition. When two
plugs are bound they can be used interchangebly in
the connect command.

The SDK profile will be generated only for the
connections of the independent plugs and plugs
that others are bound to. It means that if two
content interface plugs in an SDK are bound,
Workshop will only create one mount for the "main"
plug.