Validate binds' runtime parameters: plugs, interfaces
* Validate that all bound plugs exist and have
same interfaces when all SDKs' metadata loaded
from the store.

* Move the logic of binding for auto-connectable
plugs from creating 'connect' tasks to the the
moment when a candidate search is performed. This
way we ensure that creating of the connect tasks
is decoupled from plug binding.

* The workshop /mounts endpoint reads the actual
instance mounts to get the list instead of
traversing the SDKs' connections. Thus, decoupled
from the interface repository. It relies on the
Workshop to assign mounts to the bound plugs'
connections.

* Adding an SDK to the repo and checking the
Install policy was moved to the link-sdk handler
from auto-connect. Agent SDK installation was
moved to install-sdk -> link-sdk task set;
similarly to other SDKs to reuse the link logic
and be consistent with the flow generated for a
regular SDK. This change was required as all the
SDKs must be already in the repository by the time
auto-connect tasks start. Thus, making it possible
to cross-reference, validate and connect bound
plugs across SDKs.