Support disconnect <slot> for bound plugs
When disconnect is provided with a slot only, it
will be executed for every plug connected to a
given slot. In the case of bound plugs, disconnect
should not create a disconnect task for a bound
connection twice (once for the connection from the
resolved list and once for the connection that was
found to be bound).

This change ensures that the slot-only form is
aware of the bound plugs and disconnects them
properly.