Work around non-cancelable LXD image downloads
* The LXD backend shall still be able to cancel
the image download in progress on the task
cancellation (e.g. user aborted) and, if
restarted, come back to waiting the download
progress.

* When an image is downloaded, its aliases must be
set separately, after the operation has finished.
LXD CopyImage handles setting an image's aliases
not within the download transaction, thus, a race
is possible when two processes that initiated copy
will attempt to assign the same alias to it using
the CopyImage arguments.