Make tasks to receive download updates
* All download base image tasks must receive
accurate download updates and maintain it as their
progress. This is required for the clients to
determine the ongoing tasks and render the
progress correctly. If the download Task is in
Doing and not having proper progress updates, the
client implementation would spin the progress
spinner and not let the progress bar to render.

* If the change was aborted and restarted AND a
download is still in progress (LXD image downloads
cannot be canceled) then it must pickup the
progress of the download and still render it
correctly.