Move workshop socket to ${SNAP_COMMON}
This change implies adjusting the behaviour of:

* snap package: state files and cached SDKs are
preserved over refreshes.

* If a snap reinstalled, previous project
directories can be re-used if there is a
.worshop.lock. This required changes due to the
remove hook that removes project tracking and
workshops from LXD, so a new snap revision does
not see it. As a side effect, Workshop allows
launching a workshop per user in a project
directory if multiple users have access to it.

Also, added support of a local LXD image server
for integration and end to end tests.