Remove a connection from Repository if backend setup failed
An interface backend may fail for various reasons
(network, "source" directory does not exist for
the bind mount, etc.). The corresponding task will
be set to error and the change will be aborted.
However, the handler must also cleanup any
progress made by that time which is usually a
created connection in the interface repository.