Support interface connections in a workshop definition
* Single resource slots cannot be installed by
anyone but a host SDK (see base declarations).

* Connections that do not violate interface
declarations and passed validation via
ResolveConnect (same approach as in the case when
one connects a plug and a slot manually) will be
connected in the auto-connect task.

* Plug binding validation has moved to the
auto-connect task logic from request.go as this is
the right time and place when we have all masters
and slaves installed.

* If a plug is present in the workshop’s
connection list there will be only one candidate
pair, all other possible plugs or slots for them
are rejected. This is achieved by using a new plug
dynamic attribute "auto-explicit". Interface
declarations can differentiate between explicit
and implicit automatic connection candidates to
allow or deny an auto-connection (see the content
interface declaration).