Test workshop connections with refresh
* `workshop refresh` must preserve the connections
state and user-defined plugs & slots if failed.

* Extended the StateStorage interface with Attach
and Detach methods so the clients don't have to
use generic AddWorkshopDevice which allows for
easier fake backend implementation.

* Many apiSuite tests have changes as the daemon
creation and overlord.Loop happens in the test to
allow for the refresh use cases where one needs to
overwrite an existing workshop file in between to
initiate a proper refresh.