Fail on launch or refresh if bad plugs or slots were found
Earlier, Workshop has been sanitising bad plugs or
slots and reporting those to the corresponding
task's logs. It's not visible enough for the user
who might need to learn about the error straight
away; hence, if a plug or slot does not pass
runtime validation, the change will be aborted.