Support SDK to SDK content sharing
If content is shared between two SDKs, an internal
bind mount will be created using /etc/fstab. These
mounts will be represented as a "workshop-source"
and "workshop-target" pair. As opposed to
"host-source" and "workshop-target", when a bind
mount is created for the directory on the host.

An SDK can use a $SDK variable in
"workshop-source" to share its own content.

Renaming "host" SDK to "system" SDK will remove
confusion between a host SDK and a directory from
the host that is bound into a workshop.

Renaming content interface attributes to indicate
where source (workshop-source, host-source) and
target (workshop-target) are located will allow
for easier understanding of the `workshop info`
mounts output.