Move LXD profile logic to be part of the iface backend
Previously, the interface backend used an instance
of workshop.Backend to manipulate LXD profiles
that are created as a result of interface
connections.

This change separates the interface Backend
(lxd_device/backend.go) from the workshop backend.
It enables the backend to own and execute custom
logic required to install or remove a device of a
specific type (e.g. installing a Workshop-only
bind mount makes changes to /etc/fstab which is
not doable by only changing the LXD profile).

The interface Specification (lxd_device/spec.go)
becomes responsible for building the LXD profile
that will end up assigned to the corresponding
workshop instance.

To the client user, Workshop is an LXD instance
with profiles (one per SDK) which created when the
interfaces of its SDKs are connected.