Track revision history for an SDK
Account for multiple SDK revisions in Link and
Unlink SDK logic so that when a partial refresh
fails, the previously installed SDK revision can
be restored as a current revision. The revision
history is stored in the 'user.workshop.content'
field of a LXD configuration.

Add API tests for partial refresh.