Remove apt cache in separate lane
Removing the apt cache can fail if the workshop was created by an
earlier version, or the user manually removed the LXD volume for some
reason. This task should be in a separate lane, so the error doesn't
cause workshopd to try to undo the previous tasks (there's no way to
undo removing a workshop anyway).