Create IsProject function to replace ReadWorkshops
ReadWorkshops needs to parse project-level workshop files, which is
likely to report an error (e.g. if a user is in the middle of editing a
workshop definition), and irrelevant for the purpose of detecting
whether a directory is a project. Splitting this into a separate
function should improve separation of concerns at the cost of
introducing another function which is quite similar to ReadWorkshops.