Make file warnings opt-in when listing workshops
The workshop list command should show existing workshops and only warn
if something is wrong with the workshop definitions.

In the near future, other commands will use the same API endpoint as
workshop list. For example, workshop launch (without arguments) will use
it to detect the name of the workshop to launch. In this case it doesn't
make sense to return only the existing workshops.

To distinguish these cases, the state parameter can be either
"available" (for workshop list) or "all" (the default).