Propagate exit code from exec
Exit codes can be meaningful (e.g. for git bisect run), so we should
preserve them if possible.

I don't think we need to print an error message if the exit code is
nonzero. This should be the command's responsibility, and may not be an
error for some commands (e.g. diff or grep).