Change /v1/workshops to return workshops and files separately
Previously, GET requests would merge the list of
workshops and files presenting a file as a
workshop in an "Off" state. With the new
convention that allows multiple versions of a
workshop's definition filename, this approach is
no longer practical. workshopd is not capable of
knowing whether a workshop's definition was
renamed later on (e.g. workshop.yaml ->
.workshop/<name>.yaml). Thus, storing a file path
together with its workshop risks having some data
outdated.

Furthermore, files are not actually workshops, so
having them in a workshop list is confusing.

Therefore, it's reasonable to:
* Separate files and workshops in the API
response.
* Move the logic of matching workshops and
corresponding files to the client (matching by
name).