Add AtomicWrite method for WorkshopFs
OpenSSH 4.8 added atomic rename support as a protocol extension [1]. The
sftp library uses the old protocol by default. It does not overwrite
existing files, so cannot be used to implement atomic file writes.

The sftp library supports the atomic extension, on both the client and
server side, since version 1.13 [2]. Currently afero does not use the
extension, so I overrode Rename until that is fixed [3].

Apart from that, the implementation is modelled on `osutil.AtomicWrite`.
If more use cases arise for afero filesystems + atomics, we should
probably consolidate the implementations.

[1] https://www.openssh.com/txt/release-4.8
[2] https://github.com/pkg/sftp/pull/384
[3] https://github.com/spf13/afero/pull/437