Fix the status of an uncomplete task if aborted after a wait
The task status should be set to DoStatus if the
task is aborted after a wait. That makes the state
system to put the task on hold if, for example,
`workshop refresh --abort` was called and prevent
running the undo logic for the uncompleted task.

Previously, the task status was set to
WaitedStatus (Doing), which caused the undo logic
to run for unfinished tasks. Given that the
Workshop tasks implement revert logic to clean up
partial changes, the undo must not be run for
unfinished tasks (it's not always idempotent).

Arguably, the undo logic should not be run for
any task that wasn't in Done.