Convert errors to responses without losing context
The previous error responders used fmt.Sprintf to format error messages.
By going through fmt.Errorf instead, errors passed as arguments can be
wrapped. This will make it easier to attach an error kind to responses.

Currently errors are matched on a case-by-case basis. If this gets out
of hand, we could instead use errors.As and allow the errors to define
their own kinds via an interface method.