Store SDK meta as a volume attribute
If an SDK is installed and linked as a volume, the
meta information needs to be stored as a volume
property. This is required as Workshop uses zfs as
a storage backend which does not support access
when an instance is stopped (thus, a zfs dataset
is not mounted). Hence, commands as `info` would
fail to show mounts and SDKs info for a workshop.

The local SDKs' metadata (sketch and system), are
still loaded from the workshop filesystem.