Allow tunnel auto-connect in one direction
It should be safe to expose services running in the workshop to the
host. The other direction is not allowed, for the same reason as the
ssh-agent interface. We also avoid exposing services to other hosts.

To auto-connect a plug to a slot, they should have the same name or
appear in the connections section of the workshop definition.

Co-authored-by: Artem Konev <141050460+akcano@users.noreply.github.com>