Add tunnel connections to workshop info output
This currently uses the list of connections rather than LXD profiles,
because the system SDK profile is unavailable at the API layer. The
connections approach is a bit worse because it can't handle the case of
port 0. Port numbers inferred from the other side of the connection have
to be handled manually as well.