Introduce a resolver for workshop refresh
`workshop refresh` is going to support partial
refresh with recovery from a snapshot that
containts installation side-effects for SDKs that
were previously installed and did not receive any
updates.

This change buils a state.Change for the refresh
operation in two stages. First, it resolves
differences between a previous and a new workshop
definition and builds a refresh plan. Second, it
uses the plan to build the required task set.

No changes to the existing `workshop refresh`
behaviour except for the case when the definition
of a workshop hasn't changed and there are no new
revisions to any of its SDKs. In that case, the refresh
operation will be a no-op.