Avoid creating unix sockets outside $HOME and $XDG_RUNTIME_DIR
This is a stopgap measure to mitigate privilege escalation.
A full fix should try to create the socket as the actual user, to avoid
edge cases. The only currently known edge case is that users will
be able to overwrite files in directories that they have permission to
bind mount. This is not a huge concern since bind mounting is a
privileged operation and can already be used to bypass some checks (e.g.
mounting /home/other-user/directory can bypass execute permissions on
/home/other-user).